♻️ Refactor budget table rows into unified component
