+++ src/features/budget/components/index.ts
@@ -4,7 +4,5 @@ export { default as BudgetItem } from './BudgetItem';
 export { default as BudgetPanelHeader } from './BudgetPanelHeader';
 export { default as CategoryProgressBar } from './CategoryProgressBar';
 export { default as ExpenseTable } from './ExpenseTable';
-export { default as TableRowEdit } from './TableRowEdit';
-export { default as TableRowEntry } from './TableRowEntry';
-export { default as TableRowNew } from './TableRowNew';
+export { default as BudgetRow } from './BudgetRow';
 export { ActivitiesBudgetPopup } from './activities';
